requirements.txt
openai
python-dotenv
asyncio
httpx
fastapi
uvicorn
chainlit
